+++ md-to-pdf/tests/unit_tests/code.test.ts
@@ -1,12 +1,12 @@
 const { writeCode, writeCodeSpan } = require("../../helpers/code");
-const { getDefaultStyle, pushStyle, popStyle } = require("../../helpers/style");
+const { getDefaultStyle, pushStyle, popStyle } = require("../../helpers/styles");
 const { createMockDoc } = require("../test-utils/mockDoc");
 
 describe("md-to-pdf code helpers", () => {
   it("should write code block and increase currentHeight", () => {
     const doc = createMockDoc();
     const context = {
-      style: getDefaultStyle({ currentHeight: 70, lineDistance: 10, lineSpc: 18 }),
+      style: ({ ...getDefaultStyle(), ...{ currentHeight: 70, lineDistance: 10, lineSpc: 18 } }),
       styleStack: [],
       getDoc() {
         return doc;
@@ -43,7 +43,7 @@ describe("md-to-pdf code helpers", () => {
 
   it("should toggle code flag for inline code span", () => {
     const context = {
-      style: getDefaultStyle({ code: false }),
+      style: ({ ...getDefaultStyle(), ...{ code: false } }),
       styleStack: [],
       getStyle() {
         return this.style;
@@ -74,7 +74,7 @@ describe("md-to-pdf code helpers", () => {
   it("should apply indented code block style when codeBlockStyle is true", () => {
     const doc = createMockDoc();
     const context = {
-      style: getDefaultStyle({ currentWidth: 60, indent: 8, currentHeight: 70, lineDistance: 10, lineSpc: 18 }),
+      style: ({ ...getDefaultStyle(), ...{ currentWidth: 60, indent: 8, currentHeight: 70, lineDistance: 10, lineSpc: 18 } }),
       styleStack: [],
       getDoc() {
         return doc;
@@ -112,7 +112,7 @@ describe("md-to-pdf code helpers", () => {
   it("should handle code token without lines", () => {
     const doc = createMockDoc();
     const context = {
-      style: getDefaultStyle({ currentHeight: 70, lineDistance: 10, lineSpc: 18 }),
+      style: ({ ...getDefaultStyle(), ...{ currentHeight: 70, lineDistance: 10, lineSpc: 18 } }),
       styleStack: [],
       getDoc() {
         return doc;
