+++ package.json
@@ -27,6 +27,7 @@
   ],
   "dependencies": {
     "he": "^1.2.0",
+    "jspdf": "^4.2.1",
     "jspdf-autotable": "^5.0.2",
     "marked": "^17.0.4"
   },
